+++ src/test/java/de/tum/in/www1/artemis/exam/ExamUtilService.java
@@ -462,8 +462,8 @@ public class ExamUtilService {
         exerciseRepo.save(textExercise1);
         exerciseRepo.save(textExercise2);
 
-        QuizExercise quizExercise1 = QuizExerciseFactory.createQuizForExam(exerciseGroup1);
-        QuizExercise quizExercise2 = QuizExerciseFactory.createQuizForExam(exerciseGroup1);
+        QuizExercise quizExercise1 = quizExerciseUtilService.createQuizForExam(exerciseGroup1);
+        QuizExercise quizExercise2 = quizExerciseUtilService.createQuizForExam(exerciseGroup1);
         exerciseGroup1.setExercises(Set.of(quizExercise1, quizExercise2));
         exerciseRepo.save(quizExercise1);
         exerciseRepo.save(quizExercise2);
@@ -548,7 +548,7 @@ public class ExamUtilService {
             exam = examRepository.save(exam);
             var exerciseGroup3 = exam.getExerciseGroups().get(2 + (withProgrammingExercise ? 1 : 0));
             // Programming exercises need a proper setup for 'prepare exam start' to work
-            QuizExercise quizExercise = QuizExerciseFactory.createQuizForExam(exerciseGroup3);
+            QuizExercise quizExercise = quizExerciseUtilService.createQuizForExam(exerciseGroup3);
             exerciseRepo.save(quizExercise);
             exerciseGroup3.setExercises(Set.of(quizExercise));
         }
